backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
crewai>=0.11.0
crewai-tools>=0.0.15
langchain-groq>=0.0.1
langchain>=0.1.0
python-multipart==0.0.9
python-dotenv==1.0.1
requests
slowapi==0.1.9
